+++ src/celery_tasks/configs/config.py
@@ -1,19 +1,21 @@
+from dotenv import load_dotenv
 from pydantic import BaseSettings
 
-from enums import EnvEnum
+
+load_dotenv()
 
 
-class ServerSettings(BaseSettings):
-    APP_NAME: str = "Fast API Server"
-    APP_VERSION: str = "0.0.1"
-    APP_ENV: EnvEnum = EnvEnum.DEV
+class CeleryWorkerSettings(BaseSettings):
+    WORKER_NAME: str = "Celery Worker"
+    BROKER_URI: str = ""
+    BACKEND_URI: str = ""
 
     class Config:
         env_file = ".env"
 
 
 class ModelSettings(BaseSettings):
-    MODEL_PATH: str = "/model"
+    MODEL_PATH: str = ""
     USE_FAST_TOKENIZER: bool = True
     MODEL_MAX_LENGTH: int = 2048
 
@@ -21,5 +23,5 @@ class ModelSettings(BaseSettings):
         env_file = ".env"
 
 
-server_settings = ServerSettings()
+celery_worker_settings = CeleryWorkerSettings()
 model_settings = ModelSettings()
